whatsapp-strava-bot/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
stravalib==1.6.0
sqlalchemy==2.0.25
requests==2.31.0
python-dotenv==1.0.1
pydantic==2.5.3
pydantic-settings==2.1.0
aiohttp==3.9.1
schedule==1.2.1
python-dateutil==2.8.2
httpx==0.26.0
